pyproject.toml
[tool.poetry]
name = "lados"
version = "0.1.0"
description = "LADOS aerial oil-spill classification system"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[[tool.poetry.packages]]
include = ".."

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.6.0"
torchvision = "^0.21.0"
timm = "^0.9.0"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
celery = {extras = ["redis"], version = "^5.3.0"}
redis = "^5.0.0"
prometheus-client = "^0.19.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
pillow = "^10.1.0"
numpy = "^1.24.0"
pandas = "^2.1.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
pyyaml = "^6.0.1"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
python-multipart = "^0.0.6"
aiofiles = "^23.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
roboflow = "^1.1.0"
requests = "^2.31.0"
tqdm = "^4.66.0"
pycocotools = "^2.0.6"
onnx = "^1.15.0"
onnxruntime = "^1.16.0"
apache-airflow = "^2.7.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
ruff = "^0.1.6"
isort = "^5.12.0"
mypy = "^1.7.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

